backend_crewai_service/requirements.txt
# Python 3.14 compatibility: Using latest CrewAI 1.6.1 which may support Pydantic V2
# If issues persist, use pyenv to install Python 3.13: pyenv install 3.13.0 && pyenv local 3.13.0
Flask>=3.0.0
Flask-Cors>=4.0.0
crewai[google-genai]>=1.6.1
python-dotenv>=1.0.0
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
litellm>=1.0.0
# Force Pydantic V2 for Python 3.14 compatibility
pydantic>=2.12.0
